Added implementation of the SocketThread class
void processIncomingMessage(String message) - display the text message in the console.
void informAboutAddingNewUser(String userName) - output to the console information that a member userName has joined the chat
void informAboutDeletingNewUser(String userName) - output to the console that the participant userName has left the chat
void notifyConnectionStatusChanged(boolean clientConnected) - Set the value of the clientConnected field of the external Client object according to the passed parameter.
 Notify main thread of the Client class.